services/worker/requirements.txt
curl-cffi>=0.7.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
cryptography>=42.0.0
requests>=2.31.0
playwright>=1.43.0
patchright>=1.52.5
jwcrypto>=1.5.0
cbor2>=5.4.0
camoufox>=0.4.0
httpx>=0.27.0
selectolax>=0.3.21
names>=0.3.0
pysocks>=1.7.1
